cursor store finished, fixing tests

+++ xmtp_mls/src/groups/commit_log.rs
@@ -245,7 +245,7 @@ where
                 .get_last_cursor_for_originator(
                     &conversation.id,
                     xmtp_db::refresh_state::EntityKind::CommitLogUpload,
-                    Originators::REMOTE_COMMIT_LOG as u32,
+                    Originators::REMOTE_COMMIT_LOG,
                 )
                 .unwrap_or_default()
                 .sequence_id;
@@ -573,12 +573,12 @@ where
         let fork_check_local_cursor = conn.get_last_cursor_for_originator(
             conversation_id,
             xmtp_db::refresh_state::EntityKind::CommitLogForkCheckLocal,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )?;
         let fork_check_remote_cursor = conn.get_last_cursor_for_originator(
             conversation_id,
             xmtp_db::refresh_state::EntityKind::CommitLogForkCheckRemote,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )?;
 
         // TODO:d14n needs recheck
